refactor: :truck: move functions
